delete only if there's something

+++ workedon/workedon.py
@@ -99,7 +99,7 @@ def fetch_work(count, start_date, end_date, period, on, delete):
         with init_db():
             count = work_set.count()
             if delete:
-                if click.confirm(f"Continue deleting {count} log(s)?"):
+                if count > 0 and click.confirm(f"Continue deleting {count} log(s)?"):
                     click.echo("Deleting...")
                     deleted = Work.delete().where(Work.uuid.in_(work_set)).execute()
                     click.echo(f"{deleted} log(s) deleted successfully.")
